Remove direct call to ARNS service - this is not used.
