+++ helm_deploy/values-test.yaml
@@ -11,7 +11,6 @@ generic-service:
     SPRING_PROFILES_ACTIVE: "logstash, dev"
     APPLICATIONINSIGHTS_CONFIGURATION_FILE: applicationinsights.dev.json
     HMPPS_AUTH_BASE_URL: https://sign-in-dev.hmpps.service.justice.gov.uk
-    ARNS_API_BASE_URL: https://assess-risks-and-needs-test.hmpps.service.justice.gov.uk
     DELIUS_API_BASE_URL: https://sentence-plan-and-delius-dev.hmpps.service.justice.gov.uk
 
 # CloudPlatform AlertManager receiver to route prometheus alerts to slack
