+++ src/main/kotlin/uk/gov/justice/digital/hmpps/sentenceplan/config/WebClientConfiguration.kt
@@ -10,13 +10,9 @@ import java.time.Duration
 
 @Configuration
 class WebClientConfiguration(
-  @Value("\${arns-api.base-url}") val arnsApiBaseUrl: String,
   @Value("\${delius-api.base-url}") val deliusApiBaseUrl: String,
   @Value("\${api.timeout:20s}") val timeout: Duration,
 ) {
-  @Bean
-  fun arnsRestClient(authorizedClientManager: OAuth2AuthorizedClientManager, builder: WebClient.Builder): WebClient = builder.authorisedWebClient(authorizedClientManager, registrationId = "arns-api", url = arnsApiBaseUrl, timeout)
-
   @Bean
   fun deliusRestClient(authorizedClientManager: OAuth2AuthorizedClientManager, builder: WebClient.Builder): WebClient = builder.authorisedWebClient(authorizedClientManager, registrationId = "delius-api", url = deliusApiBaseUrl, timeout)
 }
