+++ src/main/kotlin/uk/gov/justice/digital/hmpps/sentenceplan/controller/PopInfoController.kt
@@ -7,28 +7,19 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.sentenceplan.data.CRNLinkedRequest
 import uk.gov.justice.digital.hmpps.sentenceplan.data.PopInfoResponse
-import uk.gov.justice.digital.hmpps.sentenceplan.data.RiskAssessmentResponse
-import uk.gov.justice.digital.hmpps.sentenceplan.services.ARNSApiService
+import uk.gov.justice.digital.hmpps.sentenceplan.services.DeliusApiService
 
 @RestController
 @RequestMapping("/info/pop")
 @PreAuthorize("hasAnyRole('ROLE_RISK_INTEGRATIONS_RO')")
 class PopInfoController(
-  private val arnsApiService: ARNSApiService,
+  private val deliusApiService: DeliusApiService,
 ) {
   @PostMapping
   fun getPopInfo(
     @RequestBody body: CRNLinkedRequest,
   ): PopInfoResponse {
     val crn = body.crn
-    return arnsApiService.getPopInfo(crn)
-  }
-
-  @PostMapping("/scores/risk")
-  fun getRiskScore(
-    @RequestBody body: CRNLinkedRequest,
-  ): RiskAssessmentResponse? {
-    val crn = body.crn
-    return arnsApiService.getRiskScoreInfoByCrn(crn)
+    return deliusApiService.getPopInfo(crn)
   }
 }
