+++ src/main/resources/application.yml
@@ -37,12 +37,6 @@ spring:
           hmpps-auth-service:
             token-uri: ${hmpps-auth.base-url}/auth/oauth/token
         registration:
-          arns-api:
-            provider: hmpps-auth-service
-            client-authentication-method: client_secret_basic
-            authorization-grant-type: client_credentials
-            client-id: ${api.client-id}
-            client-secret: ${api.client-secret}
           delius-api:
             provider: hmpps-auth-service
             client-authentication-method: client_secret_basic
