+++ src/test/resources/application-test.yml
@@ -8,8 +8,5 @@ management.endpoint:
 hmpps-auth:
   base-url: http://localhost:9090
 
-arns-api:
-  base-url: http://localhost:9094/arns-api
-
 delius-api:
   base-url: http://localhost:9094/delius-api
